fix(api): address CodeRabbit findings on the required-fields fill gate
1. fillByIdHandler/fillPreviewHandler moved out of their endpoint modules
   into lib/templates/fill-by-id-logic.ts and fill-preview-logic.ts
   (fillByIdLogic/fillPreviewLogic), so fill-by-id.ts and fill-preview.ts
   go back to a single default { config, handler } export, per AGENTS.md.
   assertTemplateFillUsage moved to lib/templates/template-fill-usage.ts
   so the extracted lib modules don't import from an endpoint module.
   Their tests moved alongside as fill-by-id-logic.test.ts and
   fill-preview-logic.test.ts.
2. HandlerError gained a `requiredFields` detail (mirroring the existing
   usage/confirmation/claim pattern), serialized by safeErrorBody. fill-
   by-id and fill-to-workspace now attach the full MissingRequiredField
   list instead of only a flattened message, so a client can render the
   right control per field and retry with all of them.
3. template-fill-service.test.ts's extractTexts helper now panics when a
   fixture DOCX has no word/document.xml, instead of silently reading it
   as empty text.
4. describeStoredTemplate's `arrays` grouping no longer suppresses a loop
   whose single item field happens to be named "value" — that shape is
   genuinely ambiguous between the primitive-loop convention and an
   object-item loop over `{ value }` rows, and hiding it lost the latter,
   real case entirely.
5. isMissingRequiredFieldValue now treats a non-object row inside a
   required loop item field as missing (`persons: ["invalid"]`), rather
   than silently skipping it the way mapRepeatablePath's transform callers
   do. Added findRepeatableContainer to repeatable-paths.ts so the
   required-field check can inspect raw rows itself instead of only the
   record rows mapRepeatablePath's callback would have reached.
6. The MCP fill_template/save_filled_template rejection's `issues` list
   now includes every missing required field, not just the ten-item
   message preview.

Tests added/updated for each: fill-by-id-logic.test.ts, template-fill-
service.test.ts, template-optional-defaults.test.ts, template-tools.test.ts.
